Don't us gnucore utils
Apparently, it is not a good idea to use gnucore utils on HFS and BSD.

I did run into problems with gmkdir and my notes scripts, so backing
this out.

Turns, out that all I really need is access to gdirname (because I guess
this doesn't exist on bsd) and that will set my colors for dircolors
just the way I wanted and work with zsh. So this is better. Still need
to install the coreutils.