Fix issue where certain branches don't show up in git recent
To fix this I use the right value which is the "switched to" branch. Git
on records the hash ssometimes for the left branch even if you're
switching from a branch. I not sure what cause this, but I think it
happens when I "rename" a branch, by doing a check out on something I
I just checked out.