Remove vim-snippet and vim-snippets
Could not get these to work. Not sure why. Able to get by with the snipmate
plugin which I recursivly copied in. Would like to use these in stead because
the are more modualr and seem to be more through and up todate.

Doesn't seem I can use the vim-snippets project even with snipmate even becuase
the location indexes seem to be different by one, which leads the the cursor not
being put in the correct place.

I don't mind think about snippets as a first class idea in the vim system, with
me ownining many of my snippets in the snippet folder.